Clean up ./f. Add graphs=1 option to ./f so it will just generate the graphs, open them and exit.

+++ f
@@ -1,6 +1,6 @@
 #!/bin/bash
 set -o errexit -o nounset -o pipefail
-so_targets=(
+lib_targets=(
   dakota-core
   dakota
 )
@@ -8,23 +8,32 @@ exe_targets=(
   exe-core
   exe
 )
+source_dir=/Users/robert/dakota
+build_dir=/Users/robert/dakota/zzz/build
 cat /dev/null > build.mk
-for tgt in ${so_targets[@]}; do
-  dakota-make    --var=current_source_dir=/Users/robert/dakota/$tgt \
-                 --var=source_dir=/Users/robert/dakota \
-                 --var=build_dir=/Users/robert/dakota/zzz/build \
-                 --target /Users/robert/dakota/lib/lib$tgt.dylib
-                 #--parts /Users/robert/dakota/$tgt/parts.txt
-  echo "include $tgt/build.mk" >> build.mk
+dot_files=()
+for target in ${lib_targets[@]}; do
+  dakota-make --var=current_source_dir=$source_dir/$target \
+              --var=source_dir=$source_dir \
+              --var=build_dir=$build_dir \
+              --target $source_dir/lib/lib$target.dylib
+  echo "include $target/build.mk" >> build.mk
+  dot_files+=($target/build.dot)
 done
-for tgt in ${exe_targets[@]}; do
-  dakota-make    --var=current_source_dir=/Users/robert/dakota/$tgt \
-                 --var=source_dir=/Users/robert/dakota \
-                 --var=build_dir=/Users/robert/dakota/zzz/build \
-                 --target /Users/robert/dakota/bin/$tgt
-                 #--parts /Users/robert/dakota/$tgt/parts.txt
-  echo "include $tgt/build.mk" >> build.mk
+for target in ${exe_targets[@]}; do
+  dakota-make --var=current_source_dir=$source_dir/$target \
+              --var=source_dir=$source_dir \
+              --var=build_dir=$build_dir \
+              --target $source_dir/bin/$target
+  echo "include $target/build.mk" >> build.mk
+  dot_files+=($target/build.dot)
 done
+./merge-dots.pl ${dot_files[@]} > build.dot
+graphs="${graphs:-0}"
+if [[ $graphs -ne 0 ]]; then
+  open ${dot_files[@]} build.dot
+  exit
+fi
 jobs=$(getconf _NPROCESSORS_ONLN)
 jobs=$(( jobs + 2 ))
 rm -rf zzz
